make Property internal
