+++ appenv/gen/generate.go
@@ -7,7 +7,7 @@ import (
 	"runtime"
 
 	"github.com/dave/jennifer/jen"
-	"github.com/kyoh86/gogh/appenv/prop"
+	"github.com/kyoh86/gogh/appenv/internal"
 	strcase "github.com/stoewer/go-strcase"
 )
 
@@ -64,7 +64,7 @@ func (g *Generator) createFile(packagePath string) *jen.File {
 	return file
 }
 
-func (g *Generator) parseProps(properties []*prop.Property) {
+func (g *Generator) parseProps(properties []*internal.Property) {
 	for _, p := range properties {
 		g.storeFile = g.storeFile || p.StoreFile
 		g.storeKeyring = g.storeKeyring || p.StoreKeyring
@@ -72,7 +72,7 @@ func (g *Generator) parseProps(properties []*prop.Property) {
 	}
 }
 
-func (g *Generator) doMerge(file *jen.File, properties []*prop.Property) {
+func (g *Generator) doMerge(file *jen.File, properties []*internal.Property) {
 	/* TODO:
 	- make the "Merge", LoadFile, SaveFile, LoadKeyring, SaveKeyring and GetEnvar function private
 	- create new interface Config (like "Merged").
@@ -114,13 +114,13 @@ func (g *Generator) doMerge(file *jen.File, properties []*prop.Property) {
 	})
 }
 
-func (g *Generator) tryMerge(mergeCodes *jen.Group, srcName string, p *prop.Property) {
+func (g *Generator) tryMerge(mergeCodes *jen.Group, srcName string, p *internal.Property) {
 	mergeCodes.If(jen.Id(srcName).Dot(p.Name).Op("!=").Nil()).Block(
 		jen.Id("merged").Dot(p.CamelName).Op("=").Id(srcName).Dot(p.Name).Dot("Value").Call().Assert(jen.Id(p.ValueType)),
 	)
 }
 
-func (g *Generator) doAccess(file *jen.File, properties []*prop.Property) {
+func (g *Generator) doAccess(file *jen.File, properties []*internal.Property) {
 	file.Type().Id("Accessor").StructFunc(func(accessorFields *jen.Group) {
 		if g.storeFile {
 			accessorFields.Id("file").Id("File")
@@ -188,7 +188,7 @@ func (g *Generator) doAccess(file *jen.File, properties []*prop.Property) {
 	).Line()
 }
 
-func (g *Generator) tryGet(getCodes *jen.Group, srcName string, p *prop.Property) {
+func (g *Generator) tryGet(getCodes *jen.Group, srcName string, p *internal.Property) {
 	getCodes.Block(
 		jen.Id("p").Op(":=").Id("a").Dot("parent").Dot(srcName).Dot(p.Name),
 		jen.If(jen.Id("p").Op("!=").Nil()).Block(
@@ -198,7 +198,7 @@ func (g *Generator) tryGet(getCodes *jen.Group, srcName string, p *prop.Property
 	)
 }
 
-func (g *Generator) trySet(setCodes *jen.Group, srcName string, p *prop.Property) {
+func (g *Generator) trySet(setCodes *jen.Group, srcName string, p *internal.Property) {
 	setCodes.Block(
 		jen.Id("p").Op(":=").Id("a").Dot("parent").Dot(srcName).Dot(p.Name),
 		jen.If(jen.Id("p").Op("==").Nil()).Block(
@@ -214,11 +214,11 @@ func (g *Generator) trySet(setCodes *jen.Group, srcName string, p *prop.Property
 	)
 }
 
-func (g *Generator) tryUnset(unsetCodes *jen.Group, srcName string, p *prop.Property) {
+func (g *Generator) tryUnset(unsetCodes *jen.Group, srcName string, p *internal.Property) {
 	unsetCodes.Id("a").Dot("parent").Dot(srcName).Dot(p.Name).Op("=").Nil()
 }
 
-func (g *Generator) doFile(file *jen.File, properties []*prop.Property) {
+func (g *Generator) doFile(file *jen.File, properties []*internal.Property) {
 	file.Type().Id("File").StructFunc(func(fileFields *jen.Group) {
 		for _, p := range properties {
 			if !p.StoreFile {
@@ -262,7 +262,7 @@ func (g *Generator) doFile(file *jen.File, properties []*prop.Property) {
 	file.Line()
 }
 
-func (g *Generator) doKeyring(file *jen.File, packagePath string, properties []*prop.Property) {
+func (g *Generator) doKeyring(file *jen.File, packagePath string, properties []*internal.Property) {
 	file.Type().Id("Keyring").StructFunc(func(keyringFields *jen.Group) {
 		file.Func().Id("LoadKeyring").Params().Params(jen.Id("key").Id("Keyring"), jen.Err().Id("error")).BlockFunc(func(loadKeyringCodes *jen.Group) {
 			file.Func().Id("SaveKeyring").Params(jen.Id("key").Id("Keyring")).Params(jen.Err().Id("error")).BlockFunc(func(saveKeyringCodes *jen.Group) {
@@ -307,7 +307,7 @@ func (g *Generator) doKeyring(file *jen.File, packagePath string, properties []*
 	})
 }
 
-func (g *Generator) doEnvar(file *jen.File, properties []*prop.Property) {
+func (g *Generator) doEnvar(file *jen.File, properties []*internal.Property) {
 	file.Type().Id("Envar").StructFunc(func(envarFields *jen.Group) {
 		file.Func().Id("GetEnvar").Params().Params(jen.Id("envar").Id("Envar"), jen.Err().Id("error")).BlockFunc(func(loadEnvarCodes *jen.Group) {
 			for _, p := range properties {
@@ -339,7 +339,7 @@ func (g *Generator) doEnvar(file *jen.File, properties []*prop.Property) {
 	}).Line()
 }
 
-func (g *Generator) Do(packagePath, outDir string, properties ...*prop.Property) error {
+func (g *Generator) Do(packagePath, outDir string, properties ...*internal.Property) error {
 	if err := g.init(); err != nil {
 		return err
 	}
